requirements.txt
python-dotenv~=1.0.1
rsa~=4.9
flet~=0.23.2